reformat eval, update tests
